requirements.txt
django-market
